feat: move chip truncate and tooltip logic to base component
