+++ src/components/base/StyledChip.tsx
@@ -1,9 +1,19 @@
-import { alpha, Chip, ChipProps, styled } from '@mui/material'
-import { JSX } from 'react'
+import { alpha, Chip, ChipProps, styled, Tooltip } from '@mui/material'
+import { useEffect, useRef, useState } from 'react'
 
-const StyledChip = styled(Chip)(({ theme, color = 'default' }) => {
+const BaseStyledChip = styled(Chip)(({ theme, color = 'default' }) => {
+  const baseStyles = {
+    maxWidth: '100%',
+    '& .MuiChip-label': {
+      overflow: 'hidden',
+      textOverflow: 'ellipsis',
+      whiteSpace: 'nowrap',
+    },
+  }
+
   if (color === 'default') {
     return {
+      ...baseStyles,
       color: theme.palette.text.primary,
       '& .MuiChip-icon, & .MuiChip-deleteIcon': {
         color: theme.palette.text.primary,
@@ -17,6 +27,7 @@ const StyledChip = styled(Chip)(({ theme, color = 'default' }) => {
   const palette = theme.palette[color as Exclude<typeof color, 'default'>]
 
   return {
+    ...baseStyles,
     color: palette.dark,
     '& .MuiChip-icon, & .MuiChip-deleteIcon': {
       color: palette.dark,
@@ -27,6 +38,30 @@ const StyledChip = styled(Chip)(({ theme, color = 'default' }) => {
   }
 })
 
-type PolymorphicChip = <C extends React.ElementType = 'div'>(props: ChipProps<C> & { component?: C }) => JSX.Element
+type PolymorphicChipProps<C extends React.ElementType = 'div'> = ChipProps<C> & { component?: C }
 
-export default StyledChip as PolymorphicChip
+const StyledChip = <C extends React.ElementType = 'div'>({ label, ...props }: PolymorphicChipProps<C>) => {
+  const [showTooltip, setShowTooltip] = useState(false)
+  const chipRef = useRef<HTMLDivElement>(null)
+
+  useEffect(() => {
+    if (chipRef.current && label) {
+      const labelElement = chipRef.current.querySelector('.MuiChip-label')
+      if (labelElement) {
+        setShowTooltip(labelElement.scrollWidth > labelElement.clientWidth)
+      }
+    }
+  }, [label])
+
+  const chip = <BaseStyledChip ref={chipRef} label={label} {...props} />
+
+  return showTooltip && label ? (
+    <Tooltip title={label} arrow>
+      {chip}
+    </Tooltip>
+  ) : (
+    chip
+  )
+}
+
+export default StyledChip
